Refactoring package for PiPy release.
